# Source/ACE.MarketApi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ACE.MarketApi | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "AuthEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/listings": {
      "post": {
        "tags": [
          "ListingEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "CatalogEndpoints"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seller",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dir",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/listings/suggest": {
      "get": {
        "tags": [
          "CatalogEndpoints"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionsResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/listings/{id}": {
      "get": {
        "tags": [
          "CatalogEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "account_locked",
              "attuned",
              "bad_cursor",
              "bad_limit",
              "bad_price",
              "bad_request",
              "bad_sort",
              "bad_type",
              "banned",
              "blocked_wcid",
              "busy",
              "channelling",
              "confirmation_busy",
              "confirm_timeout",
              "container_not_empty",
              "contains_attuned",
              "csrf",
              "declined",
              "gone",
              "in_trade",
              "insufficient_funds",
              "invalid_character",
              "invalid_count",
              "invalid_credentials",
              "invalid_fee",
              "invalid_price",
              "invalid_amount",
              "ip_blocked",
              "key_reused",
              "listing_limit",
              "network",
              "not_active",
              "not_available",
              "not_found",
              "not_held",
              "not_in_pack",
              "not_in_vault",
              "ok",
              "own_listing",
              "paused",
              "pet_out",
              "price_changed",
              "rate_limited",
              "recent_player_fight",
              "server",
              "trading",
              "unauthorized",
              "vault_full",
              "worn"
            ]
          },
          "price": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "BrowseResponse": {
        "required": [
          "listings",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "listings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListingResponse"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "IconLayerResponse": {
        "required": [
          "kind",
          "id",
          "url"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          },
          "url": {
            "type": "string"
          },
          "paletteTemplate": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "IconResponse": {
        "required": [
          "layers",
          "glow"
        ],
        "type": "object",
        "properties": {
          "layers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IconLayerResponse"
            }
          },
          "glow": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ListedResponse": {
        "required": [
          "id",
          "itemGuid",
          "price",
          "status",
          "listedTime"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "itemGuid": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          },
          "price": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "listedTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ListingDetailResponse": {
        "required": [
          "id",
          "itemGuid",
          "wcid",
          "name",
          "itemType",
          "material",
          "workmanship",
          "level",
          "arcaneLore",
          "summary",
          "quantity",
          "price",
          "seller",
          "listedTime",
          "wield",
          "icon",
          "lines",
          "spells"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "itemGuid": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          },
          "wcid": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          },
          "name": {
            "type": "string"
          },
          "itemType": {
            "type": "string"
          },
          "material": {
            "type": [
              "null",
              "string"
            ]
          },
          "workmanship": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "level": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "arcaneLore": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "summary": {
            "type": "string"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "price": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "seller": {
            "type": "string"
          },
          "listedTime": {
            "type": "string",
            "format": "date-time"
          },
          "wield": {
            "type": [
              "null",
              "string"
            ]
          },
          "icon": {
            "$ref": "#/components/schemas/IconResponse"
          },
          "lines": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "spells": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpellResponse"
            }
          }
        }
      },
      "ListingResponse": {
        "required": [
          "id",
          "itemGuid",
          "wcid",
          "name",
          "itemType",
          "material",
          "workmanship",
          "level",
          "arcaneLore",
          "summary",
          "quantity",
          "price",
          "seller",
          "listedTime",
          "wield",
          "icon"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "itemGuid": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          },
          "wcid": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          },
          "name": {
            "type": "string"
          },
          "itemType": {
            "type": "string"
          },
          "material": {
            "type": [
              "null",
              "string"
            ]
          },
          "workmanship": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "level": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "arcaneLore": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "summary": {
            "type": "string"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "price": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "seller": {
            "type": "string"
          },
          "listedTime": {
            "type": "string",
            "format": "date-time"
          },
          "wield": {
            "type": [
              "null",
              "string"
            ]
          },
          "icon": {
            "$ref": "#/components/schemas/IconResponse"
          }
        }
      },
      "ListRequest": {
        "required": [
          "itemGuid",
          "price",
          "characterId"
        ],
        "type": "object",
        "properties": {
          "itemGuid": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          },
          "price": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "characterId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      },
      "LoginRequest": {
        "required": [
          "account",
          "password"
        ],
        "type": "object",
        "properties": {
          "account": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "LoginResponse": {
        "required": [
          "accountId",
          "accountName"
        ],
        "type": "object",
        "properties": {
          "accountId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          },
          "accountName": {
            "type": "string"
          }
        }
      },
...
      "SpellResponse": {
        "required": [
          "name",
          "cantrip"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "cantrip": {
            "type": "boolean"
          }
        }
      },
      "SuggestionsResponse": {
        "required": [
          "suggestions"
        ],
        "type": "object",
        "properties": {
          "suggestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
...
}